+++ docs/implementation/issue-curation/issue-450-consumer-and-governance-inventory.md
@@ -68,11 +68,11 @@ in the annotation sources. Eight current bindings point at eight KlumAST classes
 
 | Annotation binding | KlumAST check implementation and direct KlumCast API use |
 |---|---|
-| `@DSL` | `CheckDslAnnotation extends KlumCastCheck<Annotation>` ([annotation](../../../klum-ast-annotations/src/main/java/com/blackbuild/groovy/configdsl/transform/DSL.java#L373-L375), [check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/validation/CheckDslAnnotation.java#L24-L39)). |
+| `@DSL` | `CheckDslAnnotation extends KlumCastCheck<Annotation>` ([annotation](../../../klum-ast-annotations/src/main/java/com/blackbuild/groovy/configdsl/transform/DSL.java#L373-L375), [check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/compiler/internal/validation/CheckDslAnnotation.java#L24-L39)). |
 | `@Field` | `FieldAstValidator extends KlumCastCheck<Annotation>` ([annotation](../../../klum-ast-annotations/src/main/java/com/blackbuild/groovy/configdsl/transform/Field.java#L56-L60), [check](../../../klum-ast/src/main/java/com/blackbuild/groovy/configdsl/transform/ast/FieldAstValidator.java#L24-L40)). |
 | `@WriteAccess` | `WriteAccessMethodCheck extends KlumCastCheck<WriteAccess>` ([annotation](../../../klum-ast-annotations/src/main/java/com/blackbuild/groovy/configdsl/transform/WriteAccess.java#L34-L38), [check](../../../klum-ast/src/main/java/com/blackbuild/groovy/configdsl/transform/ast/mutators/WriteAccessMethodCheck.java#L24-L36)). |
-| `@Validate` (two bindings) | `CheckForPrimitiveBoolean` and `ValidateAnnotationCheck`, both subclasses of `KlumCastCheck`; the latter also throws KlumCast `ValidationException` ([annotation](../../../klum-ast-annotations/src/main/java/com/blackbuild/groovy/configdsl/transform/Validate.java#L132-L136), [boolean check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/validation/CheckForPrimitiveBoolean.java#L24-L36), [placement check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/validation/ValidateAnnotationCheck.java#L24-L47)). |
-| `@Overwrite.Single` and `@Overwrite.Map` | `OverwriteSingleCheck` and `OverwriteMapCheck` extend `KlumCastCheck` ([bindings](../../../klum-ast-annotations/src/main/java/com/blackbuild/klum/ast/util/copy/Overwrite.java#L51-L76), [single check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/validation/OverwriteSingleCheck.java#L24-L42), [map check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/validation/OverwriteMapCheck.java#L24-L42)). |
+| `@Validate` (two bindings) | `CheckForPrimitiveBoolean` and `ValidateAnnotationCheck`, both subclasses of `KlumCastCheck`; the latter also throws KlumCast `ValidationException` ([annotation](../../../klum-ast-annotations/src/main/java/com/blackbuild/groovy/configdsl/transform/Validate.java#L132-L136), [boolean check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/compiler/internal/validation/CheckForPrimitiveBoolean.java#L24-L36), [placement check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/compiler/internal/validation/ValidateAnnotationCheck.java#L24-L47)). |
+| `@Overwrite.Single` and `@Overwrite.Map` | `OverwriteSingleCheck` and `OverwriteMapCheck` extend `KlumCastCheck` ([bindings](../../../klum-ast-annotations/src/main/java/com/blackbuild/klum/ast/util/copy/Overwrite.java#L51-L76), [single check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/compiler/internal/validation/OverwriteSingleCheck.java#L24-L42), [map check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/compiler/internal/validation/OverwriteMapCheck.java#L24-L42)). |
 | `@DefaultValues` | `DefaultValuesCheck extends KlumCastCheck<Annotation>` and throws KlumCast `ValidationException` for mismatched `value`/`valueTarget` contracts ([annotation](../../../klum-ast-annotations/src/main/java/com/blackbuild/klum/ast/util/layer3/annotations/DefaultValues.java#L37-L42), [check](../../../klum-ast/src/main/java/com/blackbuild/klum/ast/util/layer3/DefaultValuesCheck.java#L24-L48)). |
 
 These classes compile directly against `com.blackbuild.klum.cast.checks.impl.KlumCastCheck` and, in two cases,
