+++ klum-ast-annotations/src/main/java/com/blackbuild/groovy/configdsl/transform/DSL.java
@@ -371,7 +371,7 @@ import java.lang.annotation.Target;
         "com.blackbuild.groovy.configdsl.transform.ast.DelegatesToRWTransformation",
 })
 @KlumCastValidated
-@KlumCastValidator("com.blackbuild.klum.ast.validation.CheckDslAnnotation")
+@KlumCastValidator("com.blackbuild.klum.ast.compiler.internal.validation.CheckDslAnnotation")
 @Documented
 public @interface DSL {
     /**
