+++ klum-ast/src/main/java/com/blackbuild/klum/ast/compiler/internal/validation/OverwriteMapCheck.java
@@ -21,37 +21,31 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.blackbuild.klum.ast.validation;
+package com.blackbuild.klum.ast.compiler.internal.validation;
 
 import com.blackbuild.groovy.configdsl.transform.ast.DslAstHelper;
+import com.blackbuild.klum.ast.util.copy.OverwriteStrategy;
 import com.blackbuild.klum.cast.spi.Check;
 import com.blackbuild.klum.cast.spi.CheckContext;
 import com.blackbuild.klum.cast.spi.Diagnostic;
-import org.codehaus.groovy.ast.AnnotationNode;
 import org.codehaus.groovy.ast.ClassNode;
+import org.codehaus.groovy.ast.FieldNode;
 
 import java.util.List;
 
-import static com.blackbuild.groovy.configdsl.transform.ast.DslAstHelper.isDSLObject;
-import static com.blackbuild.klum.common.CommonAstHelper.getNullSafeClassMember;
-import static com.blackbuild.klum.common.CommonAstHelper.isAssignableTo;
+import static com.blackbuild.klum.common.CommonAstHelper.*;
 
-public class CheckDslAnnotation implements Check {
+public class OverwriteMapCheck implements Check {
+
     @Override
     public List<Diagnostic> check(CheckContext context) {
-        AnnotationNode annotationToCheck = context.getValidatedAnnotation();
-        ClassNode defaultImpl = getNullSafeClassMember(annotationToCheck, "defaultImpl", null);
-        if (defaultImpl == null) return List.of();
-        if (!isAssignableTo(defaultImpl, (ClassNode) context.getTarget()))
-            return violation(context, "defaultImpl must be a subtype of the annotated class!");
-        if (!isDSLObject(defaultImpl))
-            return violation(context, "defaultImpl must be a DSLObject!");
-        if (!DslAstHelper.isInstantiable(defaultImpl))
-            return violation(context, "defaultImpl must be instantiable!");
-        return List.of();
-    }
+        FieldNode field = (FieldNode) context.getTarget();
+        var annotationToCheck = context.getValidatedAnnotation();
+        OverwriteStrategy.Map strategy = getNullSafeEnumMemberValue(annotationToCheck, "value", OverwriteStrategy.Map.INHERIT);
+        ClassNode elementType = getElementType(field);
 
-    private List<Diagnostic> violation(CheckContext context, String message) {
-        return List.of(new Diagnostic(getClass().getName(), message, context.getValidatedAnnotation()));
+        if (strategy == OverwriteStrategy.Map.MERGE_VALUES && !DslAstHelper.isDSLObject(elementType))
+            return List.of(new Diagnostic(getClass().getName(), "MERGE_VALUES is only allowed for DSL objects", annotationToCheck));
+        return List.of();
     }
 }
